kafka/requirements.txt
kafka-python>=2.0.2
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
requests>=2.30.0
python-dateutil>=2.8.2
newsapi-python>=0.2.7
pymongo
